ASACORE-1526 Fixes to address /W4 compile issues.
Most common warnings to be resolved were:
1. Unreferenced formal parameter. Introduced UNREFERENCED_PARAMETER macros.
2. Assignment operator could not be generated - introduced private assigment operators
3. Conditional expression is constant - loop implemented like 'while(0){}' or 'while(true){}'.
Used for(;;){} or for(;;){break;} instead.
4. Conversion from 'int' to 'unit32_t'('uint8_t', 'DWORD' etc), signed/unsigned mismatch.
Introduced casts.

+++ alljoyn_core/inc/alljoyn/SimpleBusListener.h
@@ -57,7 +57,7 @@ class SimpleBusListener : public BusListener {
      * Indicator used when calling the #WaitForEvent method to indicate that the
      * listener should wait forever for the specified BusEvent.
      */
-    static const uint32_t FOREVER = -1;
+    static const uint32_t FOREVER = (uint32_t)-1;
 
     /**
      * Constructor that initializes a bus listener with specific events enabled.
